requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
schedule==1.2.2
fastapi==0.111.0
uvicorn[standard]==0.30.1
pandas==2.2.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
